chore: bump kit and migrate
