+++ course/19-adapters/package.json
@@ -1,31 +1,31 @@
 {
-  "name": "beginner-sveltekit-modules",
-  "version": "0.0.1",
-  "scripts": {
-    "dev": "vite dev",
-    "build": "svelte-kit build",
-    "package": "svelte-kit package",
-    "preview": "svelte-kit preview",
-    "prepare": "svelte-kit sync",
-    "lint": "prettier --ignore-path .gitignore --check --plugin-search-dir=. . && eslint --ignore-path .gitignore .",
-    "format": "prettier --ignore-path .gitignore --write --plugin-search-dir=. ."
-  },
-  "devDependencies": {
-    "@sveltejs/adapter-auto": "next",
-    "@sveltejs/adapter-node": "^1.0.0-next.78",
-		"@sveltejs/kit": "1.0.0-next.384",
-    "autoprefixer": "^10.4.7",
-    "eslint": "^8.12.0",
-    "eslint-config-prettier": "^8.3.0",
-    "eslint-plugin-svelte3": "^4.0.0",
-    "postcss": "^8.4.14",
-    "prettier": "^2.5.1",
-    "prettier-plugin-svelte": "^2.5.0",
-    "svelte": "^3.44.0",
-    "tailwindcss": "^3.0.24"
-  },
-  "dependencies": {
-		"vite": "^3.0.2"
+	"name": "beginner-sveltekit-modules",
+	"version": "0.0.1",
+	"scripts": {
+		"dev": "vite dev",
+		"build": "svelte-kit build",
+		"package": "svelte-kit package",
+		"preview": "svelte-kit preview",
+		"postinstall": "svelte-kit sync",
+		"lint": "prettier --ignore-path .gitignore --check --plugin-search-dir=. . && eslint --ignore-path .gitignore .",
+		"format": "prettier --ignore-path .gitignore --write --plugin-search-dir=. ."
 	},
-  "type": "module"
+	"devDependencies": {
+		"@sveltejs/adapter-auto": "next",
+		"@sveltejs/adapter-node": "^1.0.0-next.96",
+		"@sveltejs/kit": "1.0.0-next.506",
+		"autoprefixer": "^10.4.12",
+		"eslint": "^8.24.0",
+		"eslint-config-prettier": "^8.5.0",
+		"eslint-plugin-svelte3": "^4.0.0",
+		"postcss": "^8.4.17",
+		"prettier": "^2.7.1",
+		"prettier-plugin-svelte": "^2.7.1",
+		"svelte": "^3.50.1",
+		"tailwindcss": "^3.1.8"
+	},
+	"dependencies": {
+		"vite": "^3.1.4"
+	},
+	"type": "module"
 }
