+++ course/19-adapters/src/routes/(app)/+page.svelte
