+++ course/19-adapters/src/routes/(app)/product/+layout.svelte
