+++ course/19-adapters/src/routes/(app)/product/+page.svelte
