+++ course/19-adapters/src/routes/(app)/product/[name]/+server.js
@@ -1,3 +1,5 @@
+import { json } from '@sveltejs/kit';
+
 const products = [
 	{
 		name: 'cup',
@@ -27,11 +29,9 @@ const products = [
 		quantity: 1,
 		description: 'This is a sticker.'
 	}
-]
-export async function get({ params }) {
-	let product = products.find(product => product.name === params.handle)
+];
+export async function GET({ params }) {
+	let product = products.find((product) => product.name === params.name);
 
-	return {
-		body: {product}
-	};
+	return json(product);
 }
