+++ course/19-adapters/src/routes/(app)/product/apparel/+page@product.svelte
