+++ course/19-adapters/src/routes/+layout.svelte
