--- course/19-adapters/src/routes/product/__layout-product.svelte
@@ -1,4 +1,0 @@
-<div class="text3-xl font-extrabold mb-6">
-    Product
-</div>
-<slot/>
